setrt: Add more robust search for PIDs
The current search for PIDs doesn't work correctly when there is
no space at the beginning of the line so replace it with a more
robust version.  This new version first searches for the pattern
specified by the user and then searches for zero or more spaces
and then a word containing one or more numerals.

Signed-off-by: Mark Greer <mgreer@animalcreek.com>